requirements/base.txt
python-dotenv==0.15.0
websockets==8.1
